refactor: simplify auth / login to client only (for now)
