+++ packages/poolbase-app/src/app/components/Header/Header.tsx
@@ -1,15 +1,12 @@
 /** @jsx jsx */
 import { jsx, useColorMode } from 'theme-ui';
 import Router from 'next/router';
-import { PlusSquare as Add, Moon, Sun, LogOut } from 'react-feather';
+import { PlusSquare as Add, Moon, Sun } from 'react-feather';
 
-import { PropsWithAuthUser } from 'poolbase-common';
-import AppBar from 'design-system/src/components/AppBar';
-import NavIconButton from 'design-system/src/components/NavIconButton';
+import AppBar from '@poolbase/design-system/src/components/AppBar';
+import NavIconButton from '@poolbase/design-system/src/components/NavIconButton';
 
-import logout from '../../utils/auth/logout';
-
-export default function Header({ AuthUser }: PropsWithAuthUser): JSX.Element {
+export default function Header(): JSX.Element {
   const [colorMode, setColorMode] = useColorMode();
   return (
     <AppBar as="header">
@@ -20,19 +17,6 @@ export default function Header({ AuthUser }: PropsWithAuthUser): JSX.Element {
         }}
       >
         <Add />
-      </NavIconButton>
-      <NavIconButton
-        aria-label="Log out"
-        onClick={async (): Promise<void> => {
-          try {
-            await logout();
-            Router.push('/');
-          } catch (e) {
-            console.error(e);
-          }
-        }}
-      >
-        <LogOut />
       </NavIconButton>
       <NavIconButton
         onClick={(): void => setColorMode(colorMode === 'light' ? 'dark' : 'light')}
