+++ packages/poolbase-app/src/app/utils/initFirebase.ts
