--- packages/poolbase-app/src/app/utils/middleware/commonMiddleware.ts
@@ -1,6 +1,0 @@
-// Load environment variables.
-import '../../env';
-import cookieSession from './cookieSession';
-import cookieSessionRefresh from './cookieSessionRefresh';
-
-export default handler => cookieSession(cookieSessionRefresh(handler));
